+++ tests/epyccel/test_epyccel_lists.py
@@ -1,10 +1,6 @@
-# pylint: disable=missing-function-docstring, missing-module-docstring, missing-class-docstring
-# coding: utf-8
-""" Tests for list methods.
-"""
-
+# pylint: disable=missing-function-docstring, missing-module-docstring
 import pytest
-from pyccel.epyccel import epyccel
+from  pyccel.epyccel import epyccel
 
 @pytest.fixture( params=[
         pytest.param("fortran", marks = [
@@ -19,6 +15,48 @@ from pyccel.epyccel import epyccel
 )
 def language(request):
     return request.param
+
+def test_pop_last_element(language) :
+    def pop_last_element():
+        a = [1,3,45]
+        return a.pop()
+    epyc_last_element = epyccel(pop_last_element, language = language)
+    pyccel_result = epyc_last_element()
+    python_result = pop_last_element()
+    assert isinstance(python_result, type(pyccel_result))
+    assert python_result == pyccel_result
+
+
+def test_pop_specific_index(language) :
+    def pop_specific_index():
+        a = [1j,3j,45j]
+        return a.pop(1)
+    epyc_specific_index = epyccel(pop_specific_index, language = language)
+    python_result = pop_specific_index()
+    pyccel_result = epyc_specific_index()
+    assert isinstance(python_result, type(pyccel_result))
+    assert python_result == pyccel_result
+
+def test_pop_negative_index(language) :
+    def pop_negative_index():
+        a = [1j,3j,45j]
+        return a.pop(-1)
+    epyc_negative_index = epyccel(pop_negative_index, language = language)
+    python_result = pop_negative_index()
+    pyccel_result = epyc_negative_index()
+    assert isinstance(python_result, type(pyccel_result))
+    assert python_result == pyccel_result
+
+def test_pop_2(language) :
+    def pop_2():
+        a = [1.7,2.7,45.0]
+        a.pop()
+        return a.pop(-1)
+    pop_2_epyc = epyccel(pop_2, language = language)
+    python_result = pop_2()
+    pyccel_result = pop_2_epyc()
+    assert isinstance(python_result, type(pyccel_result))
+    assert python_result == pyccel_result
 
 def test_append_basic(language):
     def f():